+++ CHANGELOG.md
@@ -24,7 +24,11 @@ Also possible, but for this project less relevant, is `Deprecated` for soon-to-b
 evaluation. This is useful e.g. in afterburner calculations.
 
 ### Changed
+* Minimum supported CMake version is now version 3.28
 * The `make install` target has been standardized to create an enclosing version-specific folder inside which the standard `bin`, `lib`, `include` and `share` directories are created (previously, the version specific folder was created inside the standard ones).
+
+### Deprecated
+* CMake utility function `add_compiler_flags_if_supported` should not be used any more as it sets flags globally
 
 ## SMASH-3.3
 Date: 2025-12-03